move util functions to math.py for effect type utils
